code refactoring and fix unit tests
