+++ packages/portal-proto/src/features/cohortBuilder/style.ts
@@ -1,3 +1,5 @@
+import tw from "tailwind-styled-components";
+
 export const SecondaryTabStyle =
   "flex flex-row items-center bg-base-max text-primary border font-medium " +
   "border-solid border-primary hover:bg-secondary-darker " +
@@ -34,3 +36,27 @@ export const getCombinedClassesForRowCollapse = (
     "enabled:hover:bg-primary-darkest enabled:hover:text-primary-content-lightest enabled:hover:border-primary-darkest";
   return `${disabledClasses} ${enabledHoverClasses} ${baseClasses} ${dynamicClasses}`;
 };
+
+interface CohortGroupButtonProps {
+  disabled: boolean;
+  $isDiscard?: boolean;
+}
+
+export const CohortGroupButton = tw.button<CohortGroupButtonProps>`
+${(p: CohortGroupButtonProps) =>
+  p.disabled
+    ? "opacity-50 bg-base-max text-primary"
+    : "text-primary hover:bg-primary-darkest hover:text-primary-content-lightest bg-base-max"}
+${(p: CohortGroupButtonProps) => (p.$isDiscard ? "rounded-l" : "rounded")}
+h-12
+w-12
+flex
+justify-center
+items-center
+transition-colors
+focus-visible:outline-none
+focus-visible:ring-offset-2
+focus-visible:ring-inset
+focus-visible:ring-2
+focus-visible:ring-focusColor
+`;
